mcp-server-app/requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
flask==3.1.0
requests==2.31.0
python-dotenv==1.0.1
gunicorn==21.2.0
